fix(core): Fix import in emitter and add process.cwd to path resolution in live package
