+++ packages/indy-core/test/e2e/resources/package.json
@@ -1,20 +1,15 @@
 {
-    "name": "@jcowman/indy-test-client",
-    "private": true,
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-      "test": "mocha"
-    },
-    "author": "Joe Cowman <joe.r.cowman@gmail.com> (http://joecowman.com)",
-    "license": "MIT",
-    "dependencies": {
-      "@jcowman/indy-broken-lib": "^1.0.0"
-    },
-    "devDependencies": {
-      "chai": "^4.2.0",
-      "mocha": "^6.0.2"
-    }
+  "name": "@jcowman/indy-test-client",
+  "private": true,
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "mocha"
+  },
+  "author": "Joe Cowman <joe.r.cowman@gmail.com> (http://joecowman.com)",
+  "license": "MIT",
+  "dependencies": {
+    "@jcowman/indy-broken-lib": "^1.0.0"
   }
-  
+}
